app/requirements.txt
Flask==3.0.2
torch==2.4.0
torchvision==0.19.0
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.2.0
gunicorn==21.2.0